✨ feat: add login redirects
